Add full DAG

+++ plugins/operators/load_dimension.py
@@ -4,19 +4,25 @@ from airflow.utils.decorators import apply_defaults
 
 
 class LoadDimensionOperator(BaseOperator):
+    template_fields = ["mode"]
+    # { "mode": "delete-load" } or { "mode": "append-only" }
+
     @apply_defaults
     def __init__(self,
                  redshift_conn_id="",
                  table="",
                  sql_query="",
+                 mode="",
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.table = table
         self.sql_query = sql_query
+        self.mode = mode
 
     def execute(self, context):
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
-        redshift.run("TRUNCATE TABLE {}".format(self.table))
+        if self.mode == 'delete-load':
+            redshift.run("TRUNCATE TABLE {}".format(self.table))
         redshift.run("INSERT INTO {} ({})".format(self.table, self.sql_query))
